Tabell for inntekter

+++ src/main/kotlin/no/nav/arbeidsgiver/tiltakrefusjon/RefusjonController.kt
@@ -2,11 +2,12 @@ package no.nav.arbeidsgiver.tiltakrefusjon
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
+import java.math.BigDecimal
 
 @RestController
 class RefusjonController {
     @GetMapping("beregn")
-    fun beregn(grunnlag: Refusjonsgrunnlag): Int {
+    fun beregn(grunnlag: Refusjonsgrunnlag): BigDecimal {
         return beregnRefusjon(grunnlag)
     }
 }
